requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
openai
python-dotenv
psycopg2>=2.9.0